# subprojects/thingverse-deployment/linkerd/thingverse-api.swagger.json
{
   "swagger":"2.0",
   "info":{
      "description":"REST APIs for Thingverse. These APIs expose capability to interact with things in Thingverse.",
      "version":"0.0.1",
      "title":"Thingverse REST API Documentation",
      "termsOfService":"Terms of service",
      "contact":{
         "name":"Thingverse",
         "url":"www.thingverse.app",
         "email":"info@thingverse.app"
      },
      "license":{
         "name":"License of API",
         "url":"API license URL"
      }
   },
   "host":"localhost:9090",
   "basePath":"/",
...
   "paths":{
...
      "/api/thing":{
         "post":{
            "tags":[
               "Thing"
            ],
            "summary":"Create Thing",
            "description":"Create a Thing",
            "operationId":"createThingUsingPOST",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "*/*",
               "application/json"
            ],
            "parameters":[
               {
                  "in":"body",
                  "name":"CreateThingRequest",
                  "description":"request",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/CreateThingRequest"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/CreateThingResponse"
                  }
               },
               "201":{
                  "description":"The Thing was created",
                  "schema":{
                     "$ref":"#/definitions/CreateThingResponse"
                  }
               },
               "400":{
                  "description":"Bad requested. The server rejected your request.",
                  "schema":{
                     "$ref":"#/definitions/CreateThingResponse"
                  }
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               },
               "500":{
                  "description":"An internal error occurred.",
                  "schema":{
                     "$ref":"#/definitions/CreateThingResponse"
                  }
               }
            },
            "security":[
               {
                  "Access Token":[

                  ]
               }
            ],
            "deprecated":false
         }
      },
...
      "/api/thing/{thingID}":{
         "get":{
            "tags":[
               "Thing"
            ],
            "summary":"Get Thing",
            "description":"Get details of a Thing",
            "operationId":"getThingUsingGET",
            "produces":[
               "*/*",
               "application/json"
            ],
            "parameters":[
               {
                  "name":"thingID",
                  "in":"path",
                  "description":"thingID",
                  "required":true,
                  "type":"string"
               }
            ],
            "responses":{
               "200":{
                  "description":"The Thing was retrieved",
                  "schema":{
                     "$ref":"#/definitions/GetThingResponse"
                  }
               },
               "400":{
                  "description":"Bad requested. The server rejected your request.",
                  "schema":{
                     "$ref":"#/definitions/GetThingResponse"
                  }
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Bad requested. The thing was not found.",
                  "schema":{
                     "$ref":"#/definitions/GetThingResponse"
                  }
               },
               "500":{
                  "description":"An internal error occurred.",
                  "schema":{
                     "$ref":"#/definitions/GetThingResponse"
                  }
               }
            },
            "security":[
               {
                  "Access Token":[

                  ]
               }
            ],
            "deprecated":false
         },
         "put":{
            "tags":[
               "Thing"
            ],
            "summary":"Update Thing",
            "description":"Update a Thing",
            "operationId":"updateThingUsingPUT",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "*/*",
               "application/json"
            ],
            "parameters":[
               {
                  "in":"body",
                  "name":"UpdateThingRequest",
                  "description":"request",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/UpdateThingRequest"
                  }
               },
               {
                  "name":"thingID",
                  "in":"path",
                  "description":"thingID",
                  "required":true,
                  "type":"string"
               }
            ],
            "responses":{
               "200":{
                  "description":"The Thing was updated",
                  "schema":{
                     "$ref":"#/definitions/UpdateThingResponse"
                  }
               },
               "201":{
                  "description":"Created"
               },
               "400":{
                  "description":"Bad requested. The server rejected your request.",
                  "schema":{
                     "$ref":"#/definitions/UpdateThingResponse"
                  }
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Thing not found.",
                  "schema":{
                     "$ref":"#/definitions/UpdateThingResponse"
                  }
               },
               "500":{
                  "description":"An internal error occurred.",
                  "schema":{
                     "$ref":"#/definitions/UpdateThingResponse"
                  }
               }
            },
            "security":[
               {
                  "Access Token":[

                  ]
               }
            ],
            "deprecated":false
         }
      },
...
   },
   "securityDefinitions":{
      "Access Token":{
         "type":"apiKey",
         "name":"Authorization",
         "in":"header"
      }
   },
   "definitions":{
...
      "CreateThingRequest":{
         "type":"object",
         "properties":{
            "attributes":{
               "type":"object",
               "description":"The attribute map to be set"
            }
         },
         "title":"CreateThingRequest",
         "description":"Thing creation request."
      },
      "CreateThingResponse":{
         "type":"object",
         "properties":{
            "message":{
               "type":"string",
               "description":"Any accompanying message from the backend"
            },
            "thingID":{
               "type":"string",
               "description":"ID of the created Thing"
            }
         },
         "title":"CreateThingResponse",
         "description":"Thing creation response."
      },
...
      "GetThingResponse":{
         "type":"object",
         "properties":{
            "attributes":{
               "type":"object",
               "description":"The current attribute map"
            },
            "thingID":{
               "type":"string",
               "description":"The Thing ID"
            }
         },
         "title":"GetThingResponse",
         "description":"Thing details."
      },
...
      "UpdateThingRequest":{
         "type":"object",
         "properties":{
            "attributes":{
               "type":"object",
               "description":"The attribute map to be updated. Existing attributes are not erased."
            }
         },
         "title":"UpdateThingRequest",
         "description":"Thing updation request."
      },
      "UpdateThingResponse":{
         "type":"object",
         "properties":{
            "message":{
               "type":"string",
               "description":"A message from the backend"
            }
         },
         "title":"UpdateThingResponse",
         "description":"Thing updation response."
      },
...
   }
}